Moved maze generation code to Maze.kt
